Rename `Empty` widget to `Space`
