+++ examples/tour.rs
@@ -1,7 +1,7 @@
 use iced::{
     button, scrollable, slider, text_input, Button, Checkbox, Color, Column,
-    Container, Element, Empty, HorizontalAlignment, Image, Length, Radio, Row,
-    Sandbox, Scrollable, Settings, Slider, Text, TextInput,
+    Container, Element, HorizontalAlignment, Image, Length, Radio, Row,
+    Sandbox, Scrollable, Settings, Slider, Space, Text, TextInput,
 };
 
 pub fn main() {
@@ -67,7 +67,7 @@ impl Sandbox for Tour {
             );
         }
 
-        controls = controls.push(Empty::new().width(Length::Fill));
+        controls = controls.push(Space::with_width(Length::Fill));
 
         if steps.can_continue() {
             controls = controls.push(
