+++ src/native.rs
@@ -1,6 +1,6 @@
 pub use iced_winit::{
-    Align, Background, Color, Command, Empty, Font, HorizontalAlignment,
-    Length, Subscription, VerticalAlignment,
+    Align, Background, Color, Command, Font, HorizontalAlignment, Length,
+    Space, Subscription, VerticalAlignment,
 };
 
 pub mod widget {
