+++ web/src/widget.rs
@@ -25,10 +25,10 @@ pub mod text_input;
 mod checkbox;
 mod column;
 mod container;
-mod empty;
 mod image;
 mod radio;
 mod row;
+mod space;
 mod text;
 
 #[doc(no_inline)]
@@ -45,10 +45,10 @@ pub use text_input::TextInput;
 pub use checkbox::Checkbox;
 pub use column::Column;
 pub use container::Container;
-pub use empty::Empty;
 pub use image::Image;
 pub use radio::Radio;
 pub use row::Row;
+pub use space::Space;
 
 /// A component that displays information and allows interaction.
 ///
