+++ wgpu/src/renderer/widget.rs
@@ -1,12 +1,12 @@
 mod button;
 mod checkbox;
 mod column;
-mod empty;
 mod image;
 mod radio;
 mod row;
 mod scrollable;
 mod slider;
+mod space;
 mod text;
 mod text_input;
